all working in any OS without python 3.11
